Use Clang 0.14 generation
