+++ src/CodecZstd.jl
@@ -17,19 +17,7 @@ import TranscodingStreams:
 
 using Zstd_jll
 
-module LibZstd
-    using CEnum
-    import Zstd_jll: libzstd
-
-    const ZDICTLIB_VISIBILITY = nothing
-    const ZSTDLIB_VISIBILITY = nothing
-    const ZSTDERRORLIB_VISIBILITY = nothing
-    const INT_MAX = typemax(Cint)
-
-    include(joinpath("libzstd", "libzstd_common.jl"))
-    include(joinpath("libzstd", "libzstd_api.jl"))
-end
-
+include(joinpath("libzstd","LibZstd.jl"))
 include("libzstd.jl")
 include("compression.jl")
 include("decompression.jl")
