+++ src/libzstd/LibZstd.jl
@@ -1,17 +1,42 @@
-# Julia wrapper for header: zdict.h
-# Automatically generated using Clang.jl
+module LibZstd
+
+using Zstd_jll
+export Zstd_jll
+
+using CEnum
+
+const ZDICTLIB_VISIBILITY = nothing
+const ZSTDLIB_VISIBILITY = nothing
+const ZSTDERRORLIB_VISIBILITY = nothing
+const INT_MAX = typemax(Cint)
+
+ZSTD_LIB_VERSION = quote
+    ZSTD_VERSION_MAJOR, ZSTD_VERSION_MINOR, ZSTD_VERSION_RELEASE
+end
+
+function ZSTD_EXPAND_AND_QUOTE(expr)
+    v = eval(expr)
+    join(string.(v),".")
+end
+
 
 
 function ZDICT_trainFromBuffer(dictBuffer, dictBufferCapacity, samplesBuffer, samplesSizes, nbSamples)
-    ccall((:ZDICT_trainFromBuffer, libzstd), Csize_t, (Ptr{Cvoid}, Csize_t, Ptr{Cvoid}, Ptr{Csize_t}, UInt32), dictBuffer, dictBufferCapacity, samplesBuffer, samplesSizes, nbSamples)
+    ccall((:ZDICT_trainFromBuffer, libzstd), Csize_t, (Ptr{Cvoid}, Csize_t, Ptr{Cvoid}, Ptr{Csize_t}, Cuint), dictBuffer, dictBufferCapacity, samplesBuffer, samplesSizes, nbSamples)
+end
+
+struct ZDICT_params_t
+    compressionLevel::Cint
+    notificationLevel::Cuint
+    dictID::Cuint
 end
 
 function ZDICT_finalizeDictionary(dstDictBuffer, maxDictSize, dictContent, dictContentSize, samplesBuffer, samplesSizes, nbSamples, parameters)
-    ccall((:ZDICT_finalizeDictionary, libzstd), Csize_t, (Ptr{Cvoid}, Csize_t, Ptr{Cvoid}, Csize_t, Ptr{Cvoid}, Ptr{Csize_t}, UInt32, ZDICT_params_t), dstDictBuffer, maxDictSize, dictContent, dictContentSize, samplesBuffer, samplesSizes, nbSamples, parameters)
+    ccall((:ZDICT_finalizeDictionary, libzstd), Csize_t, (Ptr{Cvoid}, Csize_t, Ptr{Cvoid}, Csize_t, Ptr{Cvoid}, Ptr{Csize_t}, Cuint, ZDICT_params_t), dstDictBuffer, maxDictSize, dictContent, dictContentSize, samplesBuffer, samplesSizes, nbSamples, parameters)
 end
 
 function ZDICT_getDictID(dictBuffer, dictSize)
-    ccall((:ZDICT_getDictID, libzstd), UInt32, (Ptr{Cvoid}, Csize_t), dictBuffer, dictSize)
+    ccall((:ZDICT_getDictID, libzstd), Cuint, (Ptr{Cvoid}, Csize_t), dictBuffer, dictSize)
 end
 
 function ZDICT_getDictHeaderSize(dictBuffer, dictSize)
@@ -19,22 +44,19 @@ function ZDICT_getDictHeaderSize(dictBuffer, dictSize)
 end
 
 function ZDICT_isError(errorCode)
-    ccall((:ZDICT_isError, libzstd), UInt32, (Csize_t,), errorCode)
+    ccall((:ZDICT_isError, libzstd), Cuint, (Csize_t,), errorCode)
 end
 
 function ZDICT_getErrorName(errorCode)
-    ccall((:ZDICT_getErrorName, libzstd), Cstring, (Csize_t,), errorCode)
+    ccall((:ZDICT_getErrorName, libzstd), Ptr{Cchar}, (Csize_t,), errorCode)
 end
-# Julia wrapper for header: zstd.h
-# Automatically generated using Clang.jl
-
 
 function ZSTD_versionNumber()
-    ccall((:ZSTD_versionNumber, libzstd), UInt32, ())
+    ccall((:ZSTD_versionNumber, libzstd), Cuint, ())
 end
 
 function ZSTD_versionString()
-    ccall((:ZSTD_versionString, libzstd), Cstring, ())
+    ccall((:ZSTD_versionString, libzstd), Ptr{Cchar}, ())
 end
 
 function ZSTD_compress(dst, dstCapacity, src, srcSize, compressionLevel)
@@ -62,11 +84,11 @@ function ZSTD_compressBound(srcSize)
 end
 
 function ZSTD_isError(code)
-    ccall((:ZSTD_isError, libzstd), UInt32, (Csize_t,), code)
+    ccall((:ZSTD_isError, libzstd), Cuint, (Csize_t,), code)
 end
 
 function ZSTD_getErrorName(code)
-    ccall((:ZSTD_getErrorName, libzstd), Cstring, (Csize_t,), code)
+    ccall((:ZSTD_getErrorName, libzstd), Ptr{Cchar}, (Csize_t,), code)
 end
 
 function ZSTD_minCLevel()
@@ -81,6 +103,10 @@ function ZSTD_defaultCLevel()
     ccall((:ZSTD_defaultCLevel, libzstd), Cint, ())
 end
 
+mutable struct ZSTD_CCtx_s end
+
+const ZSTD_CCtx = ZSTD_CCtx_s
+
 function ZSTD_createCCtx()
     ccall((:ZSTD_createCCtx, libzstd), Ptr{ZSTD_CCtx}, ())
 end
@@ -93,6 +119,10 @@ function ZSTD_compressCCtx(cctx, dst, dstCapacity, src, srcSize, compressionLeve
     ccall((:ZSTD_compressCCtx, libzstd), Csize_t, (Ptr{ZSTD_CCtx}, Ptr{Cvoid}, Csize_t, Ptr{Cvoid}, Csize_t, Cint), cctx, dst, dstCapacity, src, srcSize, compressionLevel)
 end
 
+mutable struct ZSTD_DCtx_s end
+
+const ZSTD_DCtx = ZSTD_DCtx_s
+
 function ZSTD_createDCtx()
     ccall((:ZSTD_createDCtx, libzstd), Ptr{ZSTD_DCtx}, ())
 end
@@ -103,6 +133,61 @@ end
 
 function ZSTD_decompressDCtx(dctx, dst, dstCapacity, src, srcSize)
     ccall((:ZSTD_decompressDCtx, libzstd), Csize_t, (Ptr{ZSTD_DCtx}, Ptr{Cvoid}, Csize_t, Ptr{Cvoid}, Csize_t), dctx, dst, dstCapacity, src, srcSize)
+end
+
+@cenum ZSTD_strategy::UInt32 begin
+    ZSTD_fast = 1
+    ZSTD_dfast = 2
+    ZSTD_greedy = 3
+    ZSTD_lazy = 4
+    ZSTD_lazy2 = 5
+    ZSTD_btlazy2 = 6
+    ZSTD_btopt = 7
+    ZSTD_btultra = 8
+    ZSTD_btultra2 = 9
+end
+
+@cenum ZSTD_cParameter::UInt32 begin
+    ZSTD_c_compressionLevel = 100
+    ZSTD_c_windowLog = 101
+    ZSTD_c_hashLog = 102
+    ZSTD_c_chainLog = 103
+    ZSTD_c_searchLog = 104
+    ZSTD_c_minMatch = 105
+    ZSTD_c_targetLength = 106
+    ZSTD_c_strategy = 107
+    ZSTD_c_enableLongDistanceMatching = 160
+    ZSTD_c_ldmHashLog = 161
+    ZSTD_c_ldmMinMatch = 162
+    ZSTD_c_ldmBucketSizeLog = 163
+    ZSTD_c_ldmHashRateLog = 164
+    ZSTD_c_contentSizeFlag = 200
+    ZSTD_c_checksumFlag = 201
+    ZSTD_c_dictIDFlag = 202
+    ZSTD_c_nbWorkers = 400
+    ZSTD_c_jobSize = 401
+    ZSTD_c_overlapLog = 402
+    ZSTD_c_experimentalParam1 = 500
+    ZSTD_c_experimentalParam2 = 10
+    ZSTD_c_experimentalParam3 = 1000
+    ZSTD_c_experimentalParam4 = 1001
+    ZSTD_c_experimentalParam5 = 1002
+    ZSTD_c_experimentalParam6 = 1003
+    ZSTD_c_experimentalParam7 = 1004
+    ZSTD_c_experimentalParam8 = 1005
+    ZSTD_c_experimentalParam9 = 1006
+    ZSTD_c_experimentalParam10 = 1007
+    ZSTD_c_experimentalParam11 = 1008
+    ZSTD_c_experimentalParam12 = 1009
+    ZSTD_c_experimentalParam13 = 1010
+    ZSTD_c_experimentalParam14 = 1011
+    ZSTD_c_experimentalParam15 = 1012
+end
+
+struct ZSTD_bounds
+    error::Csize_t
+    lowerBound::Cint
+    upperBound::Cint
 end
 
 function ZSTD_cParam_getBounds(cParam)
@@ -117,12 +202,26 @@ function ZSTD_CCtx_setPledgedSrcSize(cctx, pledgedSrcSize)
     ccall((:ZSTD_CCtx_setPledgedSrcSize, libzstd), Csize_t, (Ptr{ZSTD_CCtx}, Culonglong), cctx, pledgedSrcSize)
 end
 
+@cenum ZSTD_ResetDirective::UInt32 begin
+    ZSTD_reset_session_only = 1
+    ZSTD_reset_parameters = 2
+    ZSTD_reset_session_and_parameters = 3
+end
+
 function ZSTD_CCtx_reset(cctx, reset)
     ccall((:ZSTD_CCtx_reset, libzstd), Csize_t, (Ptr{ZSTD_CCtx}, ZSTD_ResetDirective), cctx, reset)
 end
 
 function ZSTD_compress2(cctx, dst, dstCapacity, src, srcSize)
     ccall((:ZSTD_compress2, libzstd), Csize_t, (Ptr{ZSTD_CCtx}, Ptr{Cvoid}, Csize_t, Ptr{Cvoid}, Csize_t), cctx, dst, dstCapacity, src, srcSize)
+end
+
+@cenum ZSTD_dParameter::UInt32 begin
+    ZSTD_d_windowLogMax = 100
+    ZSTD_d_experimentalParam1 = 1000
+    ZSTD_d_experimentalParam2 = 1001
+    ZSTD_d_experimentalParam3 = 1002
+    ZSTD_d_experimentalParam4 = 1003
 end
 
 function ZSTD_dParam_getBounds(dParam)
@@ -137,12 +236,36 @@ function ZSTD_DCtx_reset(dctx, reset)
     ccall((:ZSTD_DCtx_reset, libzstd), Csize_t, (Ptr{ZSTD_DCtx}, ZSTD_ResetDirective), dctx, reset)
 end
 
+mutable struct ZSTD_inBuffer_s
+    src::Ptr{Cvoid}
+    size::Csize_t
+    pos::Csize_t
+end
+
+const ZSTD_inBuffer = ZSTD_inBuffer_s
+
+mutable struct ZSTD_outBuffer_s
+    dst::Ptr{Cvoid}
+    size::Csize_t
+    pos::Csize_t
+end
+
+const ZSTD_outBuffer = ZSTD_outBuffer_s
+
+const ZSTD_CStream = ZSTD_CCtx
+
 function ZSTD_createCStream()
     ccall((:ZSTD_createCStream, libzstd), Ptr{ZSTD_CStream}, ())
 end
 
 function ZSTD_freeCStream(zcs)
     ccall((:ZSTD_freeCStream, libzstd), Csize_t, (Ptr{ZSTD_CStream},), zcs)
+end
+
+@cenum ZSTD_EndDirective::UInt32 begin
+    ZSTD_e_continue = 0
+    ZSTD_e_flush = 1
+    ZSTD_e_end = 2
 end
 
 function ZSTD_compressStream2(cctx, output, input, endOp)
@@ -172,6 +295,8 @@ end
 function ZSTD_endStream(zcs, output)
     ccall((:ZSTD_endStream, libzstd), Csize_t, (Ptr{ZSTD_CStream}, Ptr{ZSTD_outBuffer}), zcs, output)
 end
+
+const ZSTD_DStream = ZSTD_DCtx
 
 function ZSTD_createDStream()
     ccall((:ZSTD_createDStream, libzstd), Ptr{ZSTD_DStream}, ())
@@ -205,6 +330,10 @@ function ZSTD_decompress_usingDict(dctx, dst, dstCapacity, src, srcSize, dict, d
     ccall((:ZSTD_decompress_usingDict, libzstd), Csize_t, (Ptr{ZSTD_DCtx}, Ptr{Cvoid}, Csize_t, Ptr{Cvoid}, Csize_t, Ptr{Cvoid}, Csize_t), dctx, dst, dstCapacity, src, srcSize, dict, dictSize)
 end
 
+mutable struct ZSTD_CDict_s end
+
+const ZSTD_CDict = ZSTD_CDict_s
+
 function ZSTD_createCDict(dictBuffer, dictSize, compressionLevel)
     ccall((:ZSTD_createCDict, libzstd), Ptr{ZSTD_CDict}, (Ptr{Cvoid}, Csize_t, Cint), dictBuffer, dictSize, compressionLevel)
 end
@@ -216,6 +345,10 @@ end
 function ZSTD_compress_usingCDict(cctx, dst, dstCapacity, src, srcSize, cdict)
     ccall((:ZSTD_compress_usingCDict, libzstd), Csize_t, (Ptr{ZSTD_CCtx}, Ptr{Cvoid}, Csize_t, Ptr{Cvoid}, Csize_t, Ptr{ZSTD_CDict}), cctx, dst, dstCapacity, src, srcSize, cdict)
 end
+
+mutable struct ZSTD_DDict_s end
+
+const ZSTD_DDict = ZSTD_DDict_s
 
 function ZSTD_createDDict(dictBuffer, dictSize)
     ccall((:ZSTD_createDDict, libzstd), Ptr{ZSTD_DDict}, (Ptr{Cvoid}, Csize_t), dictBuffer, dictSize)
@@ -230,19 +363,19 @@ function ZSTD_decompress_usingDDict(dctx, dst, dstCapacity, src, srcSize, ddict)
 end
 
 function ZSTD_getDictID_fromDict(dict, dictSize)
-    ccall((:ZSTD_getDictID_fromDict, libzstd), UInt32, (Ptr{Cvoid}, Csize_t), dict, dictSize)
+    ccall((:ZSTD_getDictID_fromDict, libzstd), Cuint, (Ptr{Cvoid}, Csize_t), dict, dictSize)
 end
 
 function ZSTD_getDictID_fromCDict(cdict)
-    ccall((:ZSTD_getDictID_fromCDict, libzstd), UInt32, (Ptr{ZSTD_CDict},), cdict)
+    ccall((:ZSTD_getDictID_fromCDict, libzstd), Cuint, (Ptr{ZSTD_CDict},), cdict)
 end
 
 function ZSTD_getDictID_fromDDict(ddict)
-    ccall((:ZSTD_getDictID_fromDDict, libzstd), UInt32, (Ptr{ZSTD_DDict},), ddict)
+    ccall((:ZSTD_getDictID_fromDDict, libzstd), Cuint, (Ptr{ZSTD_DDict},), ddict)
 end
 
 function ZSTD_getDictID_fromFrame(src, srcSize)
-    ccall((:ZSTD_getDictID_fromFrame, libzstd), UInt32, (Ptr{Cvoid}, Csize_t), src, srcSize)
+    ccall((:ZSTD_getDictID_fromFrame, libzstd), Cuint, (Ptr{Cvoid}, Csize_t), src, srcSize)
 end
 
 function ZSTD_CCtx_loadDictionary(cctx, dict, dictSize)
@@ -293,6 +426,83 @@ function ZSTD_sizeof_DDict(ddict)
     ccall((:ZSTD_sizeof_DDict, libzstd), Csize_t, (Ptr{ZSTD_DDict},), ddict)
 end
 
+mutable struct ZSTD_CCtx_params_s end
+
+const ZSTD_CCtx_params = ZSTD_CCtx_params_s
+
+struct ZSTD_Sequence
+    offset::Cuint
+    litLength::Cuint
+    matchLength::Cuint
+    rep::Cuint
+end
+
+struct ZSTD_compressionParameters
+    windowLog::Cuint
+    chainLog::Cuint
+    hashLog::Cuint
+    searchLog::Cuint
+    minMatch::Cuint
+    targetLength::Cuint
+    strategy::ZSTD_strategy
+end
+
+struct ZSTD_frameParameters
+    contentSizeFlag::Cint
+    checksumFlag::Cint
+    noDictIDFlag::Cint
+end
+
+struct ZSTD_parameters
+    cParams::ZSTD_compressionParameters
+    fParams::ZSTD_frameParameters
+end
+
+@cenum ZSTD_dictContentType_e::UInt32 begin
+    ZSTD_dct_auto = 0
+    ZSTD_dct_rawContent = 1
+    ZSTD_dct_fullDict = 2
+end
+
+@cenum ZSTD_dictLoadMethod_e::UInt32 begin
+    ZSTD_dlm_byCopy = 0
+    ZSTD_dlm_byRef = 1
+end
+
+@cenum ZSTD_format_e::UInt32 begin
+    ZSTD_f_zstd1 = 0
+    ZSTD_f_zstd1_magicless = 1
+end
+
+@cenum ZSTD_forceIgnoreChecksum_e::UInt32 begin
+    ZSTD_d_validateChecksum = 0
+    ZSTD_d_ignoreChecksum = 1
+end
+
+@cenum ZSTD_refMultipleDDicts_e::UInt32 begin
+    ZSTD_rmd_refSingleDDict = 0
+    ZSTD_rmd_refMultipleDDicts = 1
+end
+
+@cenum ZSTD_dictAttachPref_e::UInt32 begin
+    ZSTD_dictDefaultAttach = 0
+    ZSTD_dictForceAttach = 1
+    ZSTD_dictForceCopy = 2
+    ZSTD_dictForceLoad = 3
+end
+
+@cenum ZSTD_literalCompressionMode_e::UInt32 begin
+    ZSTD_lcm_auto = 0
+    ZSTD_lcm_huffman = 1
+    ZSTD_lcm_uncompressed = 2
+end
+
+@cenum ZSTD_useRowMatchFinderMode_e::UInt32 begin
+    ZSTD_urm_auto = 0
+    ZSTD_urm_disableRowMatchFinder = 1
+    ZSTD_urm_enableRowMatchFinder = 2
+end
+
 function ZSTD_findDecompressedSize(src, srcSize)
     ccall((:ZSTD_findDecompressedSize, libzstd), Culonglong, (Ptr{Cvoid}, Csize_t), src, srcSize)
 end
@@ -303,6 +513,11 @@ end
 
 function ZSTD_frameHeaderSize(src, srcSize)
     ccall((:ZSTD_frameHeaderSize, libzstd), Csize_t, (Ptr{Cvoid}, Csize_t), src, srcSize)
+end
+
+@cenum ZSTD_sequenceFormat_e::UInt32 begin
+    ZSTD_sf_noBlockDelimiters = 0
+    ZSTD_sf_explicitBlockDelimiters = 1
 end
 
 function ZSTD_generateSequences(zc, outSeqs, outSeqsSize, src, srcSize)
@@ -318,7 +533,7 @@ function ZSTD_compressSequences(cctx, dst, dstSize, inSeqs, inSeqsSize, src, src
 end
 
 function ZSTD_writeSkippableFrame(dst, dstCapacity, src, srcSize, magicVariant)
-    ccall((:ZSTD_writeSkippableFrame, libzstd), Csize_t, (Ptr{Cvoid}, Csize_t, Ptr{Cvoid}, Csize_t, UInt32), dst, dstCapacity, src, srcSize, magicVariant)
+    ccall((:ZSTD_writeSkippableFrame, libzstd), Csize_t, (Ptr{Cvoid}, Csize_t, Ptr{Cvoid}, Csize_t, Cuint), dst, dstCapacity, src, srcSize, magicVariant)
 end
 
 function ZSTD_estimateCCtxSize(compressionLevel)
@@ -393,6 +608,18 @@ function ZSTD_initStaticDDict(workspace, workspaceSize, dict, dictSize, dictLoad
     ccall((:ZSTD_initStaticDDict, libzstd), Ptr{ZSTD_DDict}, (Ptr{Cvoid}, Csize_t, Ptr{Cvoid}, Csize_t, ZSTD_dictLoadMethod_e, ZSTD_dictContentType_e), workspace, workspaceSize, dict, dictSize, dictLoadMethod, dictContentType)
 end
 
+# typedef void * ( * ZSTD_allocFunction ) ( void * opaque , size_t size )
+const ZSTD_allocFunction = Ptr{Cvoid}
+
+# typedef void ( * ZSTD_freeFunction ) ( void * opaque , void * address )
+const ZSTD_freeFunction = Ptr{Cvoid}
+
+struct ZSTD_customMem
+    customAlloc::ZSTD_allocFunction
+    customFree::ZSTD_freeFunction
+    opaque::Ptr{Cvoid}
+end
+
 function ZSTD_createCCtx_advanced(customMem)
     ccall((:ZSTD_createCCtx_advanced, libzstd), Ptr{ZSTD_CCtx}, (ZSTD_customMem,), customMem)
 end
@@ -412,6 +639,10 @@ end
 function ZSTD_createCDict_advanced(dict, dictSize, dictLoadMethod, dictContentType, cParams, customMem)
     ccall((:ZSTD_createCDict_advanced, libzstd), Ptr{ZSTD_CDict}, (Ptr{Cvoid}, Csize_t, ZSTD_dictLoadMethod_e, ZSTD_dictContentType_e, ZSTD_compressionParameters, ZSTD_customMem), dict, dictSize, dictLoadMethod, dictContentType, cParams, customMem)
 end
+
+mutable struct POOL_ctx_s end
+
+const ZSTD_threadPool = POOL_ctx_s
 
 function ZSTD_createThreadPool(numThreads)
     ccall((:ZSTD_createThreadPool, libzstd), Ptr{ZSTD_threadPool}, (Csize_t,), numThreads)
@@ -514,7 +745,7 @@ function ZSTD_compressStream2_simpleArgs(cctx, dst, dstCapacity, dstPos, src, sr
 end
 
 function ZSTD_isFrame(buffer, size)
-    ccall((:ZSTD_isFrame, libzstd), UInt32, (Ptr{Cvoid}, Csize_t), buffer, size)
+    ccall((:ZSTD_isFrame, libzstd), Cuint, (Ptr{Cvoid}, Csize_t), buffer, size)
 end
 
 function ZSTD_createDDict_byReference(dictBuffer, dictSize)
@@ -573,6 +804,15 @@ function ZSTD_resetCStream(zcs, pledgedSrcSize)
     ccall((:ZSTD_resetCStream, libzstd), Csize_t, (Ptr{ZSTD_CStream}, Culonglong), zcs, pledgedSrcSize)
 end
 
+struct ZSTD_frameProgression
+    ingested::Culonglong
+    consumed::Culonglong
+    produced::Culonglong
+    flushed::Culonglong
+    currentJobID::Cuint
+    nbActiveWorkers::Cuint
+end
+
 function ZSTD_getFrameProgression(cctx)
     ccall((:ZSTD_getFrameProgression, libzstd), ZSTD_frameProgression, (Ptr{ZSTD_CCtx},), cctx)
 end
@@ -625,6 +865,21 @@ function ZSTD_compressBegin_usingCDict_advanced(cctx, cdict, fParams, pledgedSrc
     ccall((:ZSTD_compressBegin_usingCDict_advanced, libzstd), Csize_t, (Ptr{ZSTD_CCtx}, Ptr{ZSTD_CDict}, ZSTD_frameParameters, Culonglong), cctx, cdict, fParams, pledgedSrcSize)
 end
 
+@cenum ZSTD_frameType_e::UInt32 begin
+    ZSTD_frame = 0
+    ZSTD_skippableFrame = 1
+end
+
+struct ZSTD_frameHeader
+    frameContentSize::Culonglong
+    windowSize::Culonglong
+    blockSizeMax::Cuint
+    frameType::ZSTD_frameType_e
+    headerSize::Cuint
+    dictID::Cuint
+    checksumFlag::Cuint
+end
+
 function ZSTD_getFrameHeader(zfhPtr, src, srcSize)
     ccall((:ZSTD_getFrameHeader, libzstd), Csize_t, (Ptr{ZSTD_frameHeader}, Ptr{Cvoid}, Csize_t), zfhPtr, src, srcSize)
 end
@@ -661,6 +916,15 @@ function ZSTD_copyDCtx(dctx, preparedDCtx)
     ccall((:ZSTD_copyDCtx, libzstd), Cvoid, (Ptr{ZSTD_DCtx}, Ptr{ZSTD_DCtx}), dctx, preparedDCtx)
 end
 
+@cenum ZSTD_nextInputType_e::UInt32 begin
+    ZSTDnit_frameHeader = 0
+    ZSTDnit_blockHeader = 1
+    ZSTDnit_block = 2
+    ZSTDnit_lastBlock = 3
+    ZSTDnit_checksum = 4
+    ZSTDnit_skippableFrame = 5
+end
+
 function ZSTD_nextInputType(dctx)
     ccall((:ZSTD_nextInputType, libzstd), ZSTD_nextInputType_e, (Ptr{ZSTD_DCtx},), dctx)
 end
@@ -680,14 +944,209 @@ end
 function ZSTD_insertBlock(dctx, blockStart, blockSize)
     ccall((:ZSTD_insertBlock, libzstd), Csize_t, (Ptr{ZSTD_DCtx}, Ptr{Cvoid}, Csize_t), dctx, blockStart, blockSize)
 end
-# Julia wrapper for header: zstd_errors.h
-# Automatically generated using Clang.jl
 
+@cenum ZSTD_ErrorCode::UInt32 begin
+    ZSTD_error_no_error = 0
+    ZSTD_error_GENERIC = 1
+    ZSTD_error_prefix_unknown = 10
+    ZSTD_error_version_unsupported = 12
+    ZSTD_error_frameParameter_unsupported = 14
+    ZSTD_error_frameParameter_windowTooLarge = 16
+    ZSTD_error_corruption_detected = 20
+    ZSTD_error_checksum_wrong = 22
+    ZSTD_error_dictionary_corrupted = 30
+    ZSTD_error_dictionary_wrong = 32
+    ZSTD_error_dictionaryCreation_failed = 34
+    ZSTD_error_parameter_unsupported = 40
+    ZSTD_error_parameter_outOfBound = 42
+    ZSTD_error_tableLog_tooLarge = 44
+    ZSTD_error_maxSymbolValue_tooLarge = 46
+    ZSTD_error_maxSymbolValue_tooSmall = 48
+    ZSTD_error_stage_wrong = 60
+    ZSTD_error_init_missing = 62
+    ZSTD_error_memory_allocation = 64
+    ZSTD_error_workSpace_tooSmall = 66
+    ZSTD_error_dstSize_tooSmall = 70
+    ZSTD_error_srcSize_wrong = 72
+    ZSTD_error_dstBuffer_null = 74
+    ZSTD_error_frameIndex_tooLarge = 100
+    ZSTD_error_seekableIO = 102
+    ZSTD_error_dstBuffer_wrong = 104
+    ZSTD_error_srcBuffer_wrong = 105
+    ZSTD_error_maxCode = 120
+end
 
 function ZSTD_getErrorCode(functionResult)
     ccall((:ZSTD_getErrorCode, libzstd), ZSTD_ErrorCode, (Csize_t,), functionResult)
 end
 
 function ZSTD_getErrorString(code)
-    ccall((:ZSTD_getErrorString, libzstd), Cstring, (ZSTD_ErrorCode,), code)
+    ccall((:ZSTD_getErrorString, libzstd), Ptr{Cchar}, (ZSTD_ErrorCode,), code)
 end
+
+# Skipping MacroDefinition: ZDICTLIB_VISIBILITY __attribute__ ( ( visibility ( "default" ) ) )
+
+const ZDICTLIB_API = ZDICTLIB_VISIBILITY
+
+# Skipping MacroDefinition: ZSTDLIB_VISIBILITY __attribute__ ( ( visibility ( "default" ) ) )
+
+const ZSTDLIB_API = ZSTDLIB_VISIBILITY
+
+const ZSTD_VERSION_MAJOR = 1
+
+const ZSTD_VERSION_MINOR = 5
+
+const ZSTD_VERSION_RELEASE = 0
+
+const ZSTD_VERSION_NUMBER = ZSTD_VERSION_MAJOR * 100 * 100 + ZSTD_VERSION_MINOR * 100 + ZSTD_VERSION_RELEASE
+
+# Skipping MacroDefinition: ZSTD_LIB_VERSION ZSTD_VERSION_MAJOR . ZSTD_VERSION_MINOR . ZSTD_VERSION_RELEASE
+
+const ZSTD_VERSION_STRING = ZSTD_EXPAND_AND_QUOTE(ZSTD_LIB_VERSION)
+
+const ZSTD_CLEVEL_DEFAULT = 3
+
+const ZSTD_MAGICNUMBER = 0xfd2fb528
+
+const ZSTD_MAGIC_DICTIONARY = 0xec30a437
+
+const ZSTD_MAGIC_SKIPPABLE_START = 0x184d2a50
+
+const ZSTD_MAGIC_SKIPPABLE_MASK = 0xfffffff0
+
+const ZSTD_BLOCKSIZELOG_MAX = 17
+
+const ZSTD_BLOCKSIZE_MAX = 1 << ZSTD_BLOCKSIZELOG_MAX
+
+const ZSTD_CONTENTSIZE_UNKNOWN = Culonglong(0) - 1
+
+const ZSTD_CONTENTSIZE_ERROR = Culonglong(0) - 2
+
+const ZSTD_FRAMEHEADERSIZE_MAX = 18
+
+const ZSTD_SKIPPABLEHEADERSIZE = 8
+
+const ZSTD_WINDOWLOG_MAX_32 = 30
+
+const ZSTD_WINDOWLOG_MAX_64 = 31
+
+# Skipping MacroDefinition: ZSTD_WINDOWLOG_MAX ( ( int ) ( sizeof ( size_t ) == 4 ? ZSTD_WINDOWLOG_MAX_32 : ZSTD_WINDOWLOG_MAX_64 ) )
+const ZSTD_WINDOWLOG_MAX = sizeof(Csize_t) == 4 ? ZSTD_WINDOWLOG_MAX_32 : ZSTD_WINDOWLOG_MAX_64
+
+const ZSTD_WINDOWLOG_MIN = 10
+
+const ZSTD_HASHLOG_MAX = if ZSTD_WINDOWLOG_MAX < 30
+            ZSTD_WINDOWLOG_MAX
+        else
+            30
+        end
+
+const ZSTD_HASHLOG_MIN = 6
+
+const ZSTD_CHAINLOG_MAX_32 = 29
+
+const ZSTD_CHAINLOG_MAX_64 = 30
+
+# Skipping MacroDefinition: ZSTD_CHAINLOG_MAX ( ( int ) ( sizeof ( size_t ) == 4 ? ZSTD_CHAINLOG_MAX_32 : ZSTD_CHAINLOG_MAX_64 ) )
+
+const ZSTD_CHAINLOG_MIN = ZSTD_HASHLOG_MIN
+
+const ZSTD_SEARCHLOG_MAX = ZSTD_WINDOWLOG_MAX - 1
+
+const ZSTD_SEARCHLOG_MIN = 1
+
+const ZSTD_MINMATCH_MAX = 7
+
+const ZSTD_MINMATCH_MIN = 3
+
+const ZSTD_TARGETLENGTH_MAX = ZSTD_BLOCKSIZE_MAX
+
+const ZSTD_TARGETLENGTH_MIN = 0
+
+const ZSTD_STRATEGY_MIN = ZSTD_fast
+
+const ZSTD_STRATEGY_MAX = ZSTD_btultra2
+
+const ZSTD_OVERLAPLOG_MIN = 0
+
+const ZSTD_OVERLAPLOG_MAX = 9
+
+const ZSTD_WINDOWLOG_LIMIT_DEFAULT = 27
+
+const ZSTD_LDM_HASHLOG_MIN = ZSTD_HASHLOG_MIN
+
+const ZSTD_LDM_HASHLOG_MAX = ZSTD_HASHLOG_MAX
+
+const ZSTD_LDM_MINMATCH_MIN = 4
+
+const ZSTD_LDM_MINMATCH_MAX = 4096
+
+const ZSTD_LDM_BUCKETSIZELOG_MIN = 1
+
+const ZSTD_LDM_BUCKETSIZELOG_MAX = 8
+
+const ZSTD_LDM_HASHRATELOG_MIN = 0
+
+const ZSTD_LDM_HASHRATELOG_MAX = ZSTD_WINDOWLOG_MAX - ZSTD_HASHLOG_MIN
+
+const ZSTD_TARGETCBLOCKSIZE_MIN = 64
+
+const ZSTD_TARGETCBLOCKSIZE_MAX = ZSTD_BLOCKSIZE_MAX
+
+const ZSTD_SRCSIZEHINT_MIN = 0
+
+const ZSTD_SRCSIZEHINT_MAX = INT_MAX
+
+const ZSTD_HASHLOG3_MAX = 17
+
+const ZSTD_c_rsyncable = ZSTD_c_experimentalParam1
+
+const ZSTD_c_format = ZSTD_c_experimentalParam2
+
+const ZSTD_c_forceMaxWindow = ZSTD_c_experimentalParam3
+
+const ZSTD_c_forceAttachDict = ZSTD_c_experimentalParam4
+
+const ZSTD_c_literalCompressionMode = ZSTD_c_experimentalParam5
+
+const ZSTD_c_targetCBlockSize = ZSTD_c_experimentalParam6
+
+const ZSTD_c_srcSizeHint = ZSTD_c_experimentalParam7
+
+const ZSTD_c_enableDedicatedDictSearch = ZSTD_c_experimentalParam8
+
+const ZSTD_c_stableInBuffer = ZSTD_c_experimentalParam9
+
+const ZSTD_c_stableOutBuffer = ZSTD_c_experimentalParam10
+
+const ZSTD_c_blockDelimiters = ZSTD_c_experimentalParam11
+
+const ZSTD_c_validateSequences = ZSTD_c_experimentalParam12
+
+const ZSTD_c_splitBlocks = ZSTD_c_experimentalParam13
+
+const ZSTD_c_useRowMatchFinder = ZSTD_c_experimentalParam14
+
+const ZSTD_c_deterministicRefPrefix = ZSTD_c_experimentalParam15
+
+const ZSTD_d_format = ZSTD_d_experimentalParam1
+
+const ZSTD_d_stableOutBuffer = ZSTD_d_experimentalParam2
+
+const ZSTD_d_forceIgnoreChecksum = ZSTD_d_experimentalParam3
+
+const ZSTD_d_refMultipleDDicts = ZSTD_d_experimentalParam4
+
+# Skipping MacroDefinition: ZSTDERRORLIB_VISIBILITY __attribute__ ( ( visibility ( "default" ) ) )
+
+const ZSTDERRORLIB_API = ZSTDERRORLIB_VISIBILITY
+
+# exports
+const PREFIXES = ["ZSTD_", "ZDICT_"]
+for name in names(@__MODULE__; all=true), prefix in PREFIXES
+    if startswith(string(name), prefix)
+        @eval export $name
+    end
+end
+
+end # module
